Fixed ambiguous redirect error, added new token type for it and skip the process in exec
