+++ Makefile
@@ -18,6 +18,7 @@ SRCS	:= 	$(SRC_DIR)/main.c $(SRC_DIR)/parsing/initialise.c $(SRC_DIR)/error_hand
 			$(SRC_DIR)/parsing/parse_env.c $(SRC_DIR)/parsing/tokens.c $(SRC_DIR)/parsing/quotes.c $(SRC_DIR)/parsing/tokens_two.c \
 			$(SRC_DIR)/parsing/expand_env.c $(SRC_DIR)/parsing/random_utils.c $(SRC_DIR)/parsing/syntax.c $(SRC_DIR)/parsing/syntax_two.c \
 			$(SRC_DIR)/parsing/heredoc.c $(SRC_DIR)/parsing/heredoc_syntax.c $(SRC_DIR)/parsing/quotes_two.c $(SRC_DIR)/parsing/expand_env_two.c\
+			$(SRC_DIR)/parsing/expand_env_three.c\
 			$(SRC_DIR)/exec/execute.c $(SRC_DIR)/exec/execution_branch.c $(SRC_DIR)/exec/pipe_management.c\
 			$(SRC_DIR)/exec/redirect.c $(SRC_DIR)/exec/utils.c $(SRC_DIR)/parsing/pipes.c $(SRC_DIR)/exec/exit.c \
 			$(SRC_DIR)/builtins/ft_cd.c $(SRC_DIR)/builtins/ft_echo.c \
