+++ src/exec/redirect.c
@@ -6,7 +6,7 @@
 /*   By: tcampbel <tcampbel@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/26 16:48:01 by xriera-c          #+#    #+#             */
-/*   Updated: 2024/05/28 17:18:02 by tcampbel         ###   ########.fr       */
+/*   Updated: 2024/06/05 11:39:09 by tcampbel         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
