+++ src/parsing/expand_env_two.c
@@ -6,57 +6,13 @@
 /*   By: tcampbel <tcampbel@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/30 15:01:05 by tcampbel          #+#    #+#             */
-/*   Updated: 2024/06/04 18:37:22 by tcampbel         ###   ########.fr       */
+/*   Updated: 2024/06/05 11:29:00 by tcampbel         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../inc/minishell.h"
 
-int	redirect_check(char *ptr)
-{
-	char	*str;
 
-	str = ptr;
-	if ((*str == '>' && *(str + 1) == '>')
-		|| (*str == '<' && *(str + 1) == '<'))
-		return (1);
-	else if ((*str == '>' && *(str + 1) != '>')
-		|| (*str == '<' && *(str + 1) != '<'))
-		return (1);
-	return (0);
-}
-
-
-char	*cpy_og_var(t_sh *msh, char *ptr)
-{
-	if (*ptr == '$' && *ptr)
-		msh->buffer[msh->buf_len++] = *ptr++;
-	ft_printf(2, RED":("END " $");
-	while (*ptr && *ptr != '$' && ((ft_isalnum(*ptr) == 1) || *ptr == '_'))
-	{
-		ft_putchar_fd(*ptr, 2);
-		msh->buffer[msh->buf_len++] = *ptr++;
-	}
-	ft_printf(2, ": ambiguous redirect\n");
-	return (ptr);
-}
-
-char	*amb_redirect_check(t_sh *msh, char *ptr)
-{
-	char	*str;
-
-	msh->flag = 0;
-	while ((*ptr == '<' || *ptr == '>' || ft_isspace(*ptr)) && *ptr)
-		msh->buffer[msh->buf_len++] = *ptr++;
-	str = ptr;
-	if (*ptr == '$' && ft_isspace(*(ptr + 1)) == 0 && *(ptr + 1) != '\0')
-	{
-		ptr = deref_var(msh, ptr + 1);
-		if (msh->flag == 1)
-			ptr = cpy_og_var(msh, str);
-	}
-	return (ptr);
-}
 
 void	handle_expansion(t_sh *msh, char *cmd)
 {
